+++ typings/index.d.ts
@@ -3,6 +3,18 @@ import * as mb from 'mapbox-gl';
 
 declare namespace atlas {
 
+    /**
+     * Mapbox layer options.
+     */
+    export interface MapboxLayerOptions {
+        minzoom?: number | undefined;
+        maxzoom?: number | undefined;
+        
+        filter?: any[] | undefined;
+        layout?: mb.AnyLayout | undefined;
+        paint?: mb.AnyPaint | undefined;
+    }
+
     /** A class that converts Mapbox styles into native Azure Maps classes and imports them into an Azure Maps instance. */
     export class MapboxStyleImporter {
         /**
@@ -26,9 +38,19 @@ declare namespace atlas {
         /**
          * Converts a Mapbox style layer to an Azure Map layer and adds it to a map.
          * @param mapboxLayer Mapbox layer to add to map.
-     * @param beforeLayer The ID of a layer to add this layer before.
+         * @param beforeLayer The ID of a layer to add this layer before.
+         * @returns Azure Maps layer instance or null.
          */
-        public addLayer(mapboxLayer: mb.Layer, beforeLayer?: string): void;
+        public addLayer(mapboxLayer: mb.Layer, beforeLayer?: string): azmaps.layer.Layer;
+
+        /**
+         * Updates the style options of a layer.
+         * @param layer The layer id or instance to set the options on.
+         * @param options Mapbox or Azure Maps style options to apply to the layer.
+         */
+        public setLayerOptions(layer: string | azmaps.layer.Layer, options: 
+            MapboxLayerOptions | azmaps.BubbleLayerOptions | azmaps.IconOptions | azmaps.TextOptions | azmaps.LayerOptions | azmaps.HeatMapLayerOptions | azmaps.LineLayerOptions | 
+            azmaps.PolygonLayerOptions | azmaps.PolygonExtrusionLayerOptions | azmaps.SymbolLayerOptions | azmaps.ImageLayerOptions | azmaps.TileLayerOptions): void;
     }
 }