+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-## azure-maps-mapbox-style-importer Changelog
+# azure-maps-mapbox-style-importer Changelog
+
+<a name="0.0.2"></a>
+## 0.0.2 (2021-08-30)
+
+- Added support for a combination of Azure Maps and Mapbox style options to be passed in.
+- Added `setLayerOptions` function which allows you to update the style (append) of a layer.
+- Updated `addLayer` function to return the Azure Maps layer that was created or null.
 
 <a name="0.0.1"></a>
-# 0.0.1 (2021-08-25)
+## 0.0.1 (2021-08-25)
 
-- initial release
+- Initial release.